refactor(phase-7): split planner into smbios_planner and script_renderer
