+++ src/osx_proxmox_next/script_renderer.py
@@ -1,0 +1,195 @@
+from __future__ import annotations
+
+from datetime import datetime, timezone
+from pathlib import Path
+from shlex import quote as shquote
+
+from .domain import SUPPORTED_MACOS, VmConfig
+from .smbios_planner import _sanitize_smbios
+
+__all__ = [
+    "render_script",
+    "_plist_patch_script",
+    "_build_oc_disk_script",
+]
+
+# Imported by planner for use in _opencore_steps
+_APPLE_OSK = "ourhardworkbythesewordsguardedpleasedontsteal(c)AppleComputerInc"
+
+
+def _partprobe_retry_snippet(loop_var: str) -> str:
+    """Return bash snippet that retries partprobe up to 10 times for slow storage."""
+    return (
+        f"partprobe ${loop_var} 2>/dev/null; "
+        f"for _i in $(seq 1 10); do ls ${{{loop_var}}}p* &>/dev/null && break; "
+        f"sleep 1; partprobe ${loop_var} 2>/dev/null; "
+        f"blockdev --rereadpt ${loop_var} 2>/dev/null; done"
+    )
+
+
+def render_script(config: VmConfig, steps: list) -> str:
+    meta = SUPPORTED_MACOS[config.macos]
+    now = datetime.now(timezone.utc).strftime("%Y-%m-%d %H:%M:%SZ")
+    lines = [
+        "#!/usr/bin/env bash",
+        "set -euo pipefail",
+        "",
+        f"# Generated by osx-proxmox-next on {now}",
+        f"# Target: {meta['label']} (channel={meta['channel']})",
+        f"# VMID: {config.vmid}",
+        "",
+    ]
+    for idx, step in enumerate(steps, start=1):
+        lines.append(f"echo '[{idx}/{len(steps)}] {step.title}'")
+        lines.append(step.command)
+        lines.append("")
+    return "\n".join(lines)
+
+
+def _plist_patch_script(
+    verbose_boot: bool = False,
+    is_amd: bool = False,
+    apple_services: bool = False,
+    smbios_serial: str = "",
+    smbios_uuid: str = "",
+    smbios_mlb: str = "",
+    smbios_rom: str = "",
+    smbios_model: str = "",
+) -> str:
+    """Return an inline python3 -c script that patches OpenCore's config.plist."""
+    # AMD: flip power management locks for Cascadelake-Server emulation
+    amd_patch = ""
+    if is_amd:
+        amd_patch = (
+            "kq=p[\"Kernel\"][\"Quirks\"]; "
+            "kq[\"AppleCpuPmCfgLock\"]=True; "
+            "kq[\"AppleXcpmCfgLock\"]=True; "
+        )
+
+    # PlatformInfo for Apple Services (iMessage, FaceTime, iCloud)
+    platforminfo = ""
+    if apple_services and smbios_serial:
+        s_serial = _sanitize_smbios(smbios_serial)
+        s_model = _sanitize_smbios(smbios_model, allow_comma=True)
+        s_uuid = _sanitize_smbios(smbios_uuid)
+        s_mlb = _sanitize_smbios(smbios_mlb)
+        s_rom = _sanitize_smbios(smbios_rom)
+        platforminfo = (
+            "pi=p.setdefault(\"PlatformInfo\",{}).setdefault(\"Generic\",{}); "
+            f"pi[\"SystemSerialNumber\"]=\"{s_serial}\"; "
+            f"pi[\"SystemProductName\"]=\"{s_model}\"; "
+            f"pi[\"SystemUUID\"]=\"{s_uuid}\"; "
+            f"pi[\"MLB\"]=\"{s_mlb}\"; "
+            f"pi[\"ROM\"]=bytes.fromhex(\"{s_rom}\"); "
+            "p[\"PlatformInfo\"][\"UpdateSMBIOS\"]=True; "
+            "p[\"PlatformInfo\"][\"UpdateDataHub\"]=True; "
+        )
+
+    return (
+        "python3 -c '"
+        "import plistlib; "
+        "import os; oc_dest=os.environ[\"OC_DEST\"]; "
+        "f=open(oc_dest+\"/EFI/OC/config.plist\",\"rb\"); p=plistlib.load(f); f.close(); "
+        "p[\"Misc\"][\"Security\"][\"ScanPolicy\"]=0; "
+        "p[\"Misc\"][\"Security\"][\"DmgLoading\"]=\"Any\"; "
+        "p[\"Misc\"][\"Security\"][\"SecureBootModel\"]=\"Disabled\"; "
+        "p[\"Misc\"][\"Boot\"][\"Timeout\"]=15; "
+        "p[\"Misc\"][\"Boot\"][\"PickerAttributes\"]=17; "
+        "p[\"Misc\"][\"Boot\"][\"HideAuxiliary\"]=True; "
+        "p[\"Misc\"][\"Boot\"][\"PickerMode\"]=\"External\"; "
+        "p[\"Misc\"][\"Boot\"][\"PickerVariant\"]=\"Acidanthera\\\\Syrah\"; "
+        "p[\"NVRAM\"][\"Add\"][\"7C436110-AB2A-4BBB-A880-FE41995C9F82\"][\"csr-active-config\"]=b\"\\x67\\x0f\\x00\\x00\"; "
+        f"p[\"NVRAM\"][\"Add\"][\"7C436110-AB2A-4BBB-A880-FE41995C9F82\"][\"boot-args\"]=\"keepsyms=1 debug=0x100{' -v' if verbose_boot else ''}\"; "
+        "p[\"NVRAM\"][\"Add\"][\"7C436110-AB2A-4BBB-A880-FE41995C9F82\"][\"prev-lang:kbd\"]=\"en-US:0\".encode(); "
+        "nv_del=p.setdefault(\"NVRAM\",{}).setdefault(\"Delete\",{}); "
+        "nv_del[\"7C436110-AB2A-4BBB-A880-FE41995C9F82\"]=[\"csr-active-config\",\"boot-args\",\"prev-lang:kbd\"]; "
+        "p[\"NVRAM\"][\"WriteFlash\"]=True; "
+        "p.setdefault(\"UEFI\",{}).setdefault(\"Quirks\",{})[\"RequestBootVarRouting\"]=True; "
+        "[k.update(Enabled=True) for k in p.get(\"Kernel\",{}).get(\"Add\",[]) if \"VirtualSMC\" in k.get(\"BundlePath\",\"\")]; "
+        + amd_patch
+        + platforminfo +
+        "f=open(oc_dest+\"/EFI/OC/config.plist\",\"wb\"); plistlib.dump(p,f); f.close(); "
+        "print(\"config.plist patched\")'"
+    )
+
+
+def _loop_cleanup_script(opencore_path: Path, dest: Path) -> str:
+    """Return bash snippet for loop device trap and stale cleanup."""
+    return (
+        "SRC_LOOP=''; DEST_LOOP=''; "
+        "OC_SRC=$(mktemp -d) && OC_DEST=$(mktemp -d) && export OC_SRC OC_DEST && "
+        "trap 'umount $OC_SRC 2>/dev/null; umount $OC_DEST 2>/dev/null; "
+        "[ -n \"$SRC_LOOP\" ] && losetup -d $SRC_LOOP 2>/dev/null; "
+        "[ -n \"$DEST_LOOP\" ] && losetup -d $DEST_LOOP 2>/dev/null; "
+        "rm -rf $OC_SRC $OC_DEST' EXIT; "
+        "umount $OC_SRC 2>/dev/null; umount $OC_DEST 2>/dev/null; "
+        f'for lo in $(losetup -j {shquote(str(opencore_path))} -O NAME --noheadings 2>/dev/null); do umount -l $lo* 2>/dev/null; losetup -d $lo 2>/dev/null; done; '
+        f'for lo in $(losetup -j {shquote(str(dest))} -O NAME --noheadings 2>/dev/null); do umount -l $lo* 2>/dev/null; losetup -d $lo 2>/dev/null; done; '
+    )
+
+
+def _mount_source_oc_script(opencore_path: Path) -> str:
+    """Return bash snippet to loop-mount the source OpenCore ISO."""
+    return (
+        f'SRC_LOOP=$(losetup -fP --show {shquote(str(opencore_path))}) && '
+        "{ [ -b \"$SRC_LOOP\" ] || { echo 'ERROR: losetup failed for OpenCore source ISO. Hints: modprobe loop; losetup -a; ls /dev/loop*'; false; }; } && "
+        f"{_partprobe_retry_snippet('SRC_LOOP')} && "
+        "SRC_PART=$(blkid -o device $SRC_LOOP ${SRC_LOOP}p* 2>/dev/null "
+        "| xargs -I{} sh -c 'blkid -s TYPE -o value {} 2>/dev/null | grep -q vfat && echo {}' "
+        "| head -1); "
+        "if [ -n \"$SRC_PART\" ]; then mount \"$SRC_PART\" $OC_SRC; "
+        "else echo 'WARN: No vfat partition found on source ISO via blkid, trying raw mount'; mount $SRC_LOOP $OC_SRC; fi && "
+        "{ mountpoint -q $OC_SRC || { echo \"ERROR: $OC_SRC is not mounted. Hints: file $SRC_LOOP; blkid $SRC_LOOP; dmesg | tail -5\"; false; }; } && "
+    )
+
+
+def _format_dest_oc_script(dest: Path) -> str:
+    """Return bash snippet to create, format, and mount the destination OpenCore disk."""
+    qp = shquote(str(dest))
+    return (
+        f'dd if=/dev/zero of={qp} bs=1M count=1024 && '
+        f'sgdisk -Z {qp} && '
+        f'sgdisk -n 1:0:0 -t 1:EF00 -c 1:OPENCORE {qp} && '
+        f'DEST_LOOP=$(losetup -fP --show {qp}) && '
+        "{ [ -b \"$DEST_LOOP\" ] || { echo 'ERROR: losetup failed for OpenCore destination disk. Hints: modprobe loop; losetup -a; ls /dev/loop*'; false; }; } && "
+        f"{_partprobe_retry_snippet('DEST_LOOP')} && "
+        "{ [ -b \"${DEST_LOOP}p1\" ] || { echo \"ERROR: ${DEST_LOOP}p1 not found after partprobe. Hint: Try running the script again (slow storage)\"; false; }; } && "
+        "mkfs.fat -F 32 -n OPENCORE ${DEST_LOOP}p1 && "
+        "mount ${DEST_LOOP}p1 $OC_DEST && "
+        "{ mountpoint -q $OC_DEST || { echo \"ERROR: $OC_DEST is not mounted. Hints: file ${DEST_LOOP}p1; blkid ${DEST_LOOP}p1; dmesg | tail -5\"; false; }; } && "
+    )
+
+
+def _build_oc_disk_script(
+    opencore_path: Path, recovery_path: Path, dest: Path, macos: str,
+    is_amd: bool = False, cores: int = 4, verbose_boot: bool = False,
+    apple_services: bool = False, smbios_serial: str = "",
+    smbios_uuid: str = "", smbios_mlb: str = "", smbios_rom: str = "",
+    smbios_model: str = "",
+) -> str:
+    """Build a bash script that creates a GPT+ESP OpenCore disk with patched config."""
+    plist_script = _plist_patch_script(
+        verbose_boot=verbose_boot, is_amd=is_amd,
+        apple_services=apple_services, smbios_serial=smbios_serial,
+        smbios_uuid=smbios_uuid, smbios_mlb=smbios_mlb,
+        smbios_rom=smbios_rom, smbios_model=smbios_model,
+    )
+
+    return (
+        _loop_cleanup_script(opencore_path, dest)
+        + _format_dest_oc_script(dest)
+        + _mount_source_oc_script(opencore_path)
+        # Copy OpenCore files (including hidden files)
+        + "cp -a $OC_SRC/. $OC_DEST/ && "
+        # Validate EFI structure was copied
+        "{ [ -d $OC_DEST/EFI/OC ] || { echo 'ERROR: OpenCore ISO does not contain expected EFI/OC directory. ISO may be corrupt.'; false; }; } && "
+        # Patch config.plist
+        + plist_script + " && "
+        # Fix plistlib self-closing tags that OpenCore's OcXmlLib rejects
+        "sed -i 's|<array/>|<array></array>|g; s|<dict/>|<dict></dict>|g; s|<data/>|<data></data>|g' $OC_DEST/EFI/OC/config.plist && "
+        # Hide OC partition from boot picker (shown only when user presses Space)
+        "echo Auxiliary > $OC_DEST/.contentVisibility && "
+        # Cleanup mounts (lazy unmount fallback for busy mounts)
+        "{ umount $OC_SRC || umount -l $OC_SRC; } && losetup -d $SRC_LOOP && "
+        "{ umount $OC_DEST || umount -l $OC_DEST; } && losetup -d $DEST_LOOP"
+    )
